Передаст моки основному компоненту приложения.
