+++ src/app.tsx
@@ -1,5 +1,6 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Path } from './const';
+import { Offer } from './types/offer';
 import CitiesListScreen from './pages/cities-list-screen/cities-list-screen';
 import ErrorScreen from './pages/error-screen/error-screen';
 import LoginScreen from './pages/login-screen/login-screen';
@@ -9,9 +10,10 @@ import PrivateRoute from './components/private-route';
 
 type AppScreenProps = {
   offersCount: number;
+  offers: Offer;
 }
 
-function App ({offersCount}: AppScreenProps): JSX.Element {
+function App ({offersCount, offers}: AppScreenProps): JSX.Element {
   return (
     <BrowserRouter>
       <Routes>
