+++ src/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app';
 import { Option } from './const';
+import { Offers } from './mocks/offers';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -9,6 +10,6 @@ const root = ReactDOM.createRoot(
 
 root.render(
   <React.StrictMode>
-    <App offersCount={Option.OffersCount}/>
+    <App offersCount={Option.OffersCount} offers={Offers}/>
   </React.StrictMode>
 );
