+++ src/mocks/offers.ts
@@ -1,31 +1,5 @@
 import { City, HousingType } from './mock-const';
-
-type OfferCity = City.Amsterdam | City.Brussels | City.Cologne | City.Dusseldorf | City.Hamburg | City.Paris;
-
-type OfferType = HousingType.Apartment | HousingType.Hotel | HousingType.House | HousingType.Room;
-
-type Host = {
-  avatarURL: string;
-  name: string;
-  isPro: boolean;
-}
-
-type Offer = {
-  id: string;
-  title: string;
-  description: string;
-  previewPictureURL: string;
-  pictures: string[];
-  type: OfferType;
-  city: OfferCity;
-  price: number;
-  rating: number;
-  bedroomCount: number;
-  maxGuestCount: number;
-  isPremium: boolean;
-  isFavorite: boolean;
-  host: Host;
-}
+import { Offer } from '../types/offer';
 
 const Offers: Offer[] = [
   {
